Prevent jQuery conflicts
Now reading a new version of jQuery, if necessary, into the alsLightbox
namespace and immediately releasing the global $ and jQuery variables
via noconflict().

This does mean that when the page does not load jQuery 1.6 or higher,
our script will load it, but not make the global jQuery vars available
to other elements on the page. I think this is OK.

One important caveat: the fancybox plugin applies its code directly to
the global jQuery variable. I’ve edited the version of fancybox now
included with this script to apply to alsLightbox instead.