Remove scrollbars on windows.
Issue here is that iPhone wants `overflow: scroll;` and Windows wants
`overflow: hidden;` on `.fancybox-inner`. Good news is that Fancybox is
actually fighting this out dynamically and applying the right value
inline on the `.fancybox-inner` element.

Not entirely sure why this was implemented in the first place, so be
sure to keep an eye on live assets. But i tested this on the Clinton
Foundation lightbox and it worked fine.